% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/[id]corpus.R
\name{add_terms_to_corpus}
\alias{add_terms_to_corpus}
\title{Corpus Functions}
\usage{
add_terms_to_corpus(corpus_name, new_terms,
  path_to_corpus = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/biblio-tech/CORPUS",
  local_github_repo = NULL)
}
\description{
Corpus Functions
}
