% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_key.R
\name{add_to_key}
\alias{add_to_key}
\title{Adds a new key data to the existing key as indicated by the path_to_key}
\usage{
add_to_key(path_to_key, dataframe, key_type = c("umls", "omop"))
}
\description{
Adds a new key data to the existing key as indicated by the path_to_key
}
