% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_key_words.R
\name{amend_key_words}
\alias{amend_key_words}
\title{Interactively collect keywords associated with a concept that will be used in the WHERE clause in the SQL statement}
\usage{
amend_key_words(concept_col, concept_to_amend, key_words_to_add,
  script_step_number, output_dataframe_name)
}
\arguments{
\item{concept_col}{column to use to filter for the row to amend in the dataframe captured by "script_step_number, "_", output_dataframe_name, ".RData""}

\item{concept_to_amend}{character vector of length one of the exact concept in concept_col that requires amending}

\item{key_words_to_add}{character vector of length one separated by commas of new key_words to add to the string of keywords}

\item{script_step_number}{the number assigned to the script this function is being run on. This is important because the R object will be saved with the script number as a prefix. If it is NULL, then it will not be saved.}

\item{output_dataframe_name}{character vector of length 1 that will be the name of the output data, which will be list split based on concept and additional column of the new terms}
}
\value{
dataframe with key_words with the addition of a new key_word
}
\description{
Interactively collect keywords associated with a concept that will be used in the WHERE clause in the SQL statement
}
