% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brake_if_na.R
\name{brake_if_na}
\alias{brake_if_na}
\title{Brakes if a value is NA}
\usage{
brake_if_na(value)
}
\arguments{
\item{value}{character vector of length 1}
}
\value{
warning message in the console with option to continue by pressing [ENTER]
}
\description{
Brakes if a value is NA
}
