% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brake_if_nrows_dont_match.R
\name{brake_if_nrows_dont_match}
\alias{brake_if_nrows_dont_match}
\title{Brake if nrows don't match between dataframes}
\usage{
brake_if_nrows_dont_match(dataframe1, dataframe2)
}
\arguments{
\item{dataframe1}{first dataframe}

\item{dataframe2}{second dataframe}
}
\description{
Brake if nrows don't match between dataframes
}
