% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brake_if_timestamp_not_std.R
\name{brake_if_timestamp_not_std}
\alias{brake_if_timestamp_not_std}
\title{Stops the script and asks the user to press enter if a timestamp isn't in standard format (ie 2019-11-21 08:00:00)}
\usage{
brake_if_timestamp_not_std(vector)
}
\arguments{
\item{vector}{timestamp vector}
}
\value{
interactive message
}
\description{
Stops the script and asks the user to press enter if a timestamp isn't in standard format (ie 2019-11-21 08:00:00)
}
