% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartograph_key.R
\name{cartograph_key}
\alias{cartograph_key}
\title{Takes a key or dataframe joined with a key and divides it based on status mapped out in a cartograph}
\usage{
cartograph_key(dataframe)
}
\arguments{
\item{dataframe}{dataframe with the necessary columns mapped out in the function}
}
\description{
Takes a key or dataframe joined with a key and divides it based on status mapped out in a cartograph
}
