% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_key_words.R
\name{collect_key_words}
\alias{collect_key_words}
\title{Interactively collect keywords associated with a concept that will be used in the WHERE clause in the SQL statement}
\usage{
collect_key_words(dataframe, script_step_number, output_dataframe_name)
}
\arguments{
\item{dataframe}{dataframe the column of source concepts that need to be mapped}

\item{script_step_number}{the number assigned to the script this function is being run on. This is important because the R object will be saved with the script number as a prefix. If it is NULL, then it will not be saved.}

\item{output_dataframe_name}{character vector of length 1 that will be the name of the output data, which will be list split based on concept and additional column of the new terms}
}
\value{
original dataframe with a KEY_WORD column as a comma-separated string
}
\description{
Interactively collect keywords associated with a concept that will be used in the WHERE clause in the SQL statement
}
