% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_redcap_parsed_data.R
\name{create_redcap_parsed_data}
\alias{create_redcap_parsed_data}
\title{Parses source redcap data dictionary to IDENTITY format}
\usage{
create_redcap_parsed_data(redcap_source_file)
}
\arguments{
\item{redcap_source_file}{path to csv export of REDCap data dictionary}
}
\value{
identity format of the source redcap data dictionary, where there are 3 additional columns for the parsed permissible values, an ID column, and a timestamp
}
\description{
Parses source redcap data dictionary to IDENTITY format
}
