% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_umls_strs_in_athena.R
\name{find_all_umls_strs_in_athena}
\alias{find_all_umls_strs_in_athena}
\title{This function takes all the possible strings associated with a UMLS cui, splits each individual string into tokens and performs a "LIKE" query on the combination of tokens per string in athena}
\usage{
find_all_umls_strs_in_athena(input_df, cui_col, label_col)
}
\arguments{
\item{input_df}{Input dataframe should have a minimum of 2 columns. One of the CUI and the 2nd of the label for the cui, which can be either an exact string associated with the CUI or another label}
}
\description{
This function takes all the possible strings associated with a UMLS cui, splits each individual string into tokens and performs a "LIKE" query on the combination of tokens per string in athena
}
