% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bibliotech_paths.R
\name{find_bibliotech_paths}
\alias{find_bibliotech_paths}
\title{Gets all the full file paths in biblio-tech for function calls}
\usage{
find_bibliotech_paths(path_to_bibliotech_repo, parts = c("CATALOGUE", "KEY"))
}
\arguments{
\item{parts}{which parts of the biblio-tech that will be included in the object returned}
}
\value{
list of length of `parts`, non-recursive file list and dir list
}
\description{
Gets all the full file paths in biblio-tech for function calls
}
