% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_source_redcap_data_dictionary.R
\name{format_source_redcap_data_dictionary}
\alias{format_source_redcap_data_dictionary}
\title{Parses source redcap data dictionary to IDENTITY format}
\usage{
format_source_redcap_data_dictionary(path_to_redcap_data_dictionary_csv,
  project_alias, identity_id_starting_digit, identity_id_prefix)
}
\arguments{
\item{project_alias}{character string of length 1 for "PROJECT_ALIAS"}

\item{identity_id_starting_digit}{character string of length 1 for the number the primary key should start at}

\item{identity_id_prefix}{character string of length 1 for the prefix each primary key should have}
}
\value{
identity format of the source redcap data dictionary, where there are 3 additional columns for the parsed permissible values, an ID column, and a timestamp
}
\description{
Parses source redcap data dictionary to IDENTITY format
}
