% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_ask_athena.R
\name{funnel_ask_athena}
\alias{funnel_ask_athena}
\title{Query based on search terms the funnels in the order it is inputted and that does not write to catalogue}
\usage{
funnel_ask_athena(...)
}
\arguments{
\item{...}{vector of phrases to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms the funnels in the order it is inputted and that does not write to catalogue
}
