% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_var_pvs_in_docx.R
\name{list_var_pvs_in_docx}
\alias{list_var_pvs_in_docx}
\title{Lists variables with each unique pv in a dataframe to a docx with given output_fn}
\usage{
list_var_pvs_in_docx(
  data,
  document_title = "",
  document_subtitle = "",
  output_fn
)
}
\description{
Lists variables with each unique pv in a dataframe to a docx with given output_fn
}
