% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_filter_key_for_current_version.R
\name{load_and_filter_key_for_current_version}
\alias{load_and_filter_key_for_current_version}
\title{Gets KEY data and cleans up based on parameters}
\usage{
load_and_filter_key_for_current_version(path_to_key, log = TRUE)
}
\arguments{
\item{path_to_key}{full file path to key csv file}

\item{log}{TRUE if this event will be logged using the log functions from the mirCat package}
}
\value{
dataframe read from path_to_key that is grouped by the IDENTITY_ID, KEY_FIELD, and KEY_CONCEPT_NAME and filtered for the most recent record
}
\description{
Gets KEY data and cleans up based on parameters
}
