% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_key.R
\name{load_key}
\alias{load_key}
\title{Gets KEY data and cleans up based on parameters}
\usage{
load_key(path_to_key, log = TRUE, project_alias)
}
\arguments{
\item{path_to_key}{full file path to key csv file}

\item{log}{TRUE if this event will be logged using the log functions from the mirCat package}

\item{project_alias}{since the KEY has all elements of all projects, project_alias is used to subset for the project}
}
\value{
dataframe read from path_to_key that is grouped by the IDENTITY_ID, KEY_FIELD, and KEY_CONCEPT_NAME and filtered for the most recent record
}
\description{
Gets KEY data and cleans up based on parameters
}
