% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosely_ask_misterconso.R
\name{loosely_ask_misterconso}
\alias{loosely_ask_misterconso}
\title{Query based on search terms that does not write to catalogue}
\usage{
loosely_ask_misterconso(...)
}
\description{
Query based on search terms that does not write to catalogue
}
