% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosely_ask_radlex.R
\name{loosely_ask_radlex}
\alias{loosely_ask_radlex}
\title{Queries RadLex.xls as if it is a database}
\usage{
loosely_ask_radlex(
  phrase,
 
    path_to_radlex_xls = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/biblio-tech/CATALOGUE/RadLex_v4/Radlex.xls"
)
}
\description{
Queries RadLex.xls as if it is a database
}
