% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Converts numbers to numbers as words}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{character vector of numbers}
}
\description{
Converts numbers to numbers as words
}
