% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_new_redcap_dd.R
\name{parse_new_redcap_dd}
\alias{parse_new_redcap_dd}
\title{Parses a new redcap data dictionary}
\usage{
parse_new_redcap_dd(
  path_to_redcap_dd,
  project_alias,
  identity_id_starting_digit,
  identity_id_prefix
)
}
\description{
Parses a new redcap data dictionary
}
