% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_redcap_boolean_values.R
\name{parse_redcap_boolean_values}
\alias{parse_redcap_boolean_values}
\title{Parses 'yesno' field types as permissible values}
\usage{
parse_redcap_boolean_values(
  dataframe,
  id_col,
  variable_col,
  field_type_col,
  new_permissible_col_name = "PERMISSIBLE_VALUE_LITERAL"
)
}
\description{
Parses 'yesno' field types as permissible values
}
