% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_redcap_permissible_values.R
\name{parse_redcap_permissible_values}
\alias{parse_redcap_permissible_values}
\title{Parses REDCap Permissible Values}
\usage{
parse_redcap_permissible_values(dataframe, id_col, variable_col,
  permissible_value_string_col,
  new_permissible_col_name = "PERMISSIBLE_VALUE_LITERAL")
}
\description{
Parses REDCap Permissible Values
}
