% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_source_redcap_data_dictionary.R
\name{parse_source_redcap_data_dictionary}
\alias{parse_source_redcap_data_dictionary}
\title{Parses source version of REDCAP data dictionary into variables and permissible values}
\usage{
parse_source_redcap_data_dictionary(input.file, write_excel_file = TRUE)
}
\arguments{
\item{input.file}{source REDCap data dictionary as csv}

\item{write_excel_file}{TRUE if an excel output with the same basename with "_parsed" appended is desired.}
}
\value{
list of dataframes for each Variables, Permissible Values, and Data Type Map
}
\description{
Parses source version of REDCAP data dictionary into variables and permissible values
}
