% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_new_redcap_data_dictionary.R
\name{process_new_redcap_data_dictionary}
\alias{process_new_redcap_data_dictionary}
\title{Process new redcap data dictionary into the parsed, IDENTITY format and added to the IDENTITY component and biblio-tech repo is commited}
\usage{
process_new_redcap_data_dictionary(path_to_redcap_data_dictionary_csv,
  project_alias, identity_id_starting_digit, identity_id_prefix,
  commit_message)
}
\arguments{
\item{project_alias}{character string of length 1 for "PROJECT_ALIAS"}

\item{identity_id_starting_digit}{character string of length 1 for the number the primary key should start at}

\item{identity_id_prefix}{character string of length 1 for the prefix each primary key should have}

\item{commit_message}{commit message when new changes are pushed the biblio-tech origin master}
}
\description{
Process new redcap data dictionary into the parsed, IDENTITY format and added to the IDENTITY component and biblio-tech repo is commited
}
