% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cui_in_umls.R
\name{query_cui_in_umls}
\alias{query_cui_in_umls}
\title{Queries cui in UMLS in MRCONSO table of metathesaurus}
\usage{
query_cui_in_umls(umls_cui, limit = NULL)
}
\arguments{
\item{umls_cui}{C-prefixed CUI from metathesaurus}

\item{limit}{LIMIT in the sql statement. If NULL, all records will be returned.}
}
\value{
resultset from SQL query
}
\description{
Queries cui in UMLS in MRCONSO table of metathesaurus
}
