% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_key.R
\name{read_key}
\alias{read_key}
\title{Reads key}
\usage{
read_key(path_to_key, project_alias, log = TRUE)
}
\arguments{
\item{project_alias}{project_alias for the specific project in the key; "all" if all projects desired.}
}
\description{
Reads key
}
