% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_athena_net.R
\name{return_athena_net}
\alias{return_athena_net}
\title{Writes Athena CONCEPT table NET}
\usage{
return_athena_net(phrases,
  path_to_athena_lookup = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/biblio-tech/CATALOGUE/Athena_Vocabulary_v5/LOOKUP.csv",
  path_to_net_dir = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/biblio-tech/CATALOGUE/Athena_Vocabulary_v5/NETS",
  sub_forward_slash = TRUE, trim_inner_space = TRUE,
  sub_underscore = TRUE, id_start_index = 60000, path_to_repo = NULL,
  return_lookup = FALSE)
}
\arguments{
\item{phrases}{character vector of length 1 or greater of queries}

\item{path_to_athena_lookup}{path to lookup table}

\item{id_start_index}{if number of rows of the lookup is 0, the starting index for the id number assigned}

\item{path_to_repo}{if not NULL, will add, commit, and push each iterative change to origin master of the repo}
}
\value{
csv files for each unique phrase queried in concept table and at the conclusion of the interation, returns the most updated Athena lookup to the console
}
\description{
Writes Athena CONCEPT table NET
}
