% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_mrconso_net.R
\name{return_mrconso_net}
\alias{return_mrconso_net}
\title{Returns MRCONSO NETS that match the phrase}
\usage{
return_mrconso_net(
  phrase,
  path_to_lookup,
 
    path_to_umls_net_dir = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/TerminologyBiblioTech/CATALOGUE/UMLS/NETS",
  sub_forward_slash = TRUE,
  trim_inner_space = TRUE,
  sub_underscore = TRUE,
  push_to_github = TRUE,
  cast_wide_net = TRUE,
  list_by_cui = TRUE
)
}
\arguments{
\item{phrase}{character vector of length 1.}

\item{cast_wide_net}{TRUE if any NET with a grep logical match is to be included}

\item{list_by_cui}{selects only CUI and STR variables and splits output by CUI into a list}

\item{ward_slash}{forward slashes throws errors with the write_csv and read_csv proceses downstream and are removed from the phrase by default}
}
\description{
Returns MRCONSO NETS that match the phrase
}
