% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_omop_domain_ids.R
\name{return_omop_domain_ids}
\alias{return_omop_domain_ids}
\title{Returns all possible OMOP domains in Athena}
\usage{
return_omop_domain_ids()
}
\value{
dataframe of unique `domain_id` variables from the `concept` table
}
\description{
Returns all possible OMOP domains in Athena
}
