% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_key.R
\name{search_key}
\alias{search_key}
\title{Search key for a term that is not specific and requires input of the key path}
\usage{
search_key(path_to_key, string)
}
\arguments{
\item{path_to_key}{path to key file}

\item{string}{string to search}
}
\description{
Search key for a term that is not specific and requires input of the key path
}
