% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_key_as_local_library.R
\name{search_key_as_local_library}
\alias{search_key_as_local_library}
\title{Searches local library concepts in key by only searching within variables marked with a "LOCAL" prefix}
\usage{
search_key_as_local_library(path_to_key, string, type = c("like", "exact"))
}
\arguments{
\item{path_to_key}{path to key file. Key must contain local concepts columns.}

\item{string}{string to search}

\item{type}{whether to search an exact match for the string or not.}
}
\description{
Searches local library concepts in key by only searching within variables marked with a "LOCAL" prefix
}
