% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_omop_key.R
\name{search_omop_key}
\alias{search_omop_key}
\title{Search OMOP key for a term}
\usage{
search_omop_key(
  path_to_key = "/Users/patelm9/GitHub/MSK/biblio-tech/KEY/REDCap/KEY_TO_OMOP.csv",
  string,
  type = c("like", "exact")
)
}
\arguments{
\item{path_to_key}{path to key file}

\item{string}{string to search}
}
\description{
Search OMOP key for a term
}
