% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_umls_key.R
\name{search_umls_key}
\alias{search_umls_key}
\title{Search umls key for a term}
\usage{
search_umls_key(
  path_to_key = "/Users/patelm9/GitHub/MSK/biblio-tech/KEY/REDCap/KEY_TO_UMLS.csv",
  string,
  type = c("like", "exact")
)
}
\arguments{
\item{path_to_key}{path to key file}

\item{string}{string to search}
}
\description{
Search umls key for a term
}
