% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_key.R
\name{split_key}
\alias{split_key}
\title{Splits KEY into PRECOORDINATED and ONE_TO_ONE or VARIABLE and PERMISSIBLE_VALUE}
\usage{
split_key(key_df, by = c("mapping ratio", "key field"))
}
\arguments{
\item{key_df}{KEY$CURRENT or KEY$HISTORY to split}

\item{by}{"mapping ratio" if a list of PRECOORDINATED vs ONE_TO_ONE is desired. "key field" if a variable vs. permissible_value split is desired.}
}
\value{
list of key_df split into 2 dataframes
}
\description{
Splits KEY into PRECOORDINATED and ONE_TO_ONE or VARIABLE and PERMISSIBLE_VALUE
}
