% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strictly_ask_athena.R
\name{strictly_ask_athena}
\alias{strictly_ask_athena}
\title{Query based on search terms that does not write to catalogue}
\usage{
strictly_ask_athena(phrase, limit = 1)
}
\arguments{
\item{...}{vector of phrases to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms that does not write to catalogue
}
