% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strictly_ask_misterconso.R
\name{strictly_ask_misterconso}
\alias{strictly_ask_misterconso}
\title{Query based on search terms}
\usage{
strictly_ask_misterconso(phrase, limit = 1)
}
\description{
Query based on search terms
}
