% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimis.R
\name{trimis}
\alias{trimis}
\title{Trims extra spaces within strings}
\usage{
trimis(vector)
}
\arguments{
\item{vector}{character vector of length one or greater}
}
\value{
vector with inner whitespace removed
}
\description{
Trims extra spaces within strings
}
