% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{umls_glossary_stt_desc}
\alias{umls_glossary_stt_desc}
\title{Definitions of TTY and STT abbreviations in MRCONSO of UMLS Metathesaurus}
\format{dataframe}
\source{
\url{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html#TTYC}
}
\usage{
data(umls_glossary_stt_desc)
}
\description{
A list of 3 dataframes that provide descriptions of the TTY and STT variable abbreviations from the 2019 version of UMLS Metathesaurus
}
